will ADD: feature of being transported to newly created admin page upon successful form entry. and also some admin form use cases polished
