--- solar-panel-planner/src/AdminPage.jsx
@@ -1,9 +1,0 @@
-function AdminPage() {
-  return (
-    <section className="h-screen flex items-center justify-center bg-green-100">
-      <h1 className="text-5xl text-green-700">Welcome, Admin!</h1>
-    </section>
-  );
-}
-
-export default AdminPage;
